New package for team sharing feature
