+++ server/server.go
@@ -39,6 +39,7 @@ import (
 	"github.com/trackit/trackit-server/routes"
 	_ "github.com/trackit/trackit-server/s3/costs"
 	_ "github.com/trackit/trackit-server/users"
+	_ "github.com/trackit/trackit-server/users/shared_account"
 	_ "github.com/trackit/trackit-server/costs/tags"
 )
